feat: Implement global footer
