--- src/composition/Footer.tsx
@@ -1,33 +1,0 @@
-import { DevNotice } from "components/Footer/DevNotice";
-import { FooterLinks } from "components/Footer/FooterLinks";
-import { TechStack } from "components/Footer/TechStack";
-import { Divider } from "design-system/Divider";
-import { Flex } from "design-system/Flex";
-import { PageFooter } from "design-system/Footer";
-import { Text } from "design-system/Text";
-
-export const Footer = () => {
-  return (
-    <PageFooter style={{ contentVisibility: "auto" }}>
-      <section>
-        <Divider className="mt-0 mb-8" />
-
-        <Text as="h2" $fontSize="1.25rem">
-          icyJoseph
-        </Text>
-
-        <DevNotice />
-
-        <Flex justifyContent="space-between" gap="2rem">
-          <TechStack />
-
-          <FooterLinks />
-        </Flex>
-
-        <Text mx="auto" pt={5} $textAlign="center">
-          &#169; {new Date().getFullYear()} icyJoseph AB - Stockholm, Sweden
-        </Text>
-      </section>
-    </PageFooter>
-  );
-};
